+++ utilities/Globals.h
@@ -46,7 +46,7 @@ struct global_settings {
     TCamera pCamera; // parametry kamery
     TCamera pDebugCamera;
     std::array<glm::dvec3, 10> FreeCameraInit; // pozycje kamery
-    std::array<glm::dvec3, 10> FreeCameraInitAngle;
+    std::array<glm::vec3, 10> FreeCameraInitAngle;
     int iCameraLast{ -1 };
     int iSlowMotion{ 0 }; // info o malym FPS: 0-OK, 1-wyłączyć multisampling, 3-promień 1.5km, 7-1km
     basic_light DayLight;
